refactor(api-service): remove common transport re-exports
