--- src/services/apiService/common/compatHeaders.ts
@@ -1,1 +1,0 @@
-export { buildCompatUserIdHeaders } from "~/services/apiTransport/compatHeaders"
