--- src/services/apiService/common/constant.ts
@@ -1,4 +1,0 @@
-// Temporary compatibility re-export for callers still importing the legacy
-// apiService path. New code should import from `~/services/apiTransport/constant`.
-// Remove this alias once the remaining legacy imports are migrated.
-export { REQUEST_CONFIG } from "~/services/apiTransport/constant"
