--- src/services/apiService/common/errors.ts
@@ -1,6 +1,0 @@
-export {
-  API_ERROR_CODES,
-  ApiError,
-  isTempWindowUnsupportedErrorCode,
-  type ApiErrorCode,
-} from "~/services/apiTransport/errors"
