--- src/services/apiService/common/minIntervalLimiter.ts
@@ -1,1 +1,0 @@
-export { createMinIntervalLimiter } from "~/services/apiTransport/minIntervalLimiter"
