--- src/services/apiService/common/siteRequestLimiter.ts
@@ -1,4 +1,0 @@
-export {
-  createSiteRequestLimiter,
-  withSiteApiRequestLimit,
-} from "~/services/apiTransport/siteRequestLimiter"
