+++ src/services/productAnalytics/actions.ts
@@ -1,6 +1,6 @@
 import { createLogger } from "~/utils/core/logger"
 
-import { API_ERROR_CODES } from "../apiService/common/errors"
+import { API_ERROR_CODES } from "../apiTransport/errors"
 import type { ProductAnalyticsActionContext } from "./actionConfig"
 import {
   PRODUCT_ANALYTICS_ERROR_CATEGORIES,
