+++ tests/services/apiService/newApiFamily/accountRefresh.test.ts
@@ -1,11 +1,11 @@
 import { beforeEach, describe, expect, it, vi } from "vitest"
 
-import { API_ERROR_CODES, ApiError } from "~/services/apiService/common/errors"
 import {
   defaultAccountRefreshImplementation,
   refreshAccountData,
   validateAccountConnection,
 } from "~/services/apiService/newApiFamily/default/accountRefresh"
+import { API_ERROR_CODES, ApiError } from "~/services/apiTransport/errors"
 import {
   AuthTypeEnum,
   SiteHealthStatus,
