+++ tests/services/apiService/newApiFamily/channelManagement.test.ts
@@ -1,6 +1,5 @@
 import { beforeEach, describe, expect, it, vi } from "vitest"
 
-import { ApiError } from "~/services/apiService/common/errors"
 import {
   createChannel,
   deleteChannel,
@@ -11,6 +10,7 @@ import {
   updateChannelModelMapping,
   updateChannelModels,
 } from "~/services/apiService/newApiFamily/channelManagement"
+import { ApiError } from "~/services/apiTransport/errors"
 import { AuthTypeEnum } from "~/types"
 
 const { mockFetchApi, mockFetchApiData } = vi.hoisted(() => ({
