+++ tests/services/apiService/sub2api/index.test.ts
@@ -3,7 +3,6 @@ import { beforeEach, describe, expect, it, vi } from "vitest"
 import { ACCOUNT_BROWSER_SESSION_SOURCES } from "~/services/accountBrowserSession/types"
 import type { ApiServiceAccountRequest } from "~/services/accounts/accountDataModel"
 import type { CreateTokenRequest } from "~/services/accountTokens/tokenProvisioningModel"
-import { API_ERROR_CODES, ApiError } from "~/services/apiService/common/errors"
 import { fetchApi } from "~/services/apiService/common/utils"
 import {
   createApiToken,
@@ -46,6 +45,7 @@ import type {
   Sub2ApiAnnouncementListData,
   Sub2ApiEnvelope,
 } from "~/services/apiService/sub2api/type"
+import { API_ERROR_CODES, ApiError } from "~/services/apiTransport/errors"
 import { AuthTypeEnum, SiteHealthStatus } from "~/types"
 
 const { mockGetLatestAuth, mockPersistAuthUpdate } = vi.hoisted(() => ({
