+++ tests/services/detectSiteType.fallback.test.ts
@@ -1,7 +1,7 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest"
 
 import { SITE_TYPES } from "~/constants/siteType"
-import { ApiError } from "~/services/apiService/common/errors"
+import { ApiError } from "~/services/apiTransport/errors"
 import {
   fetchSiteOriginalTitle,
   getAccountSiteType,
