+++ tests/services/modelList/accountSources/tokenScopedFallback.test.ts
@@ -1,7 +1,7 @@
 import { beforeEach, describe, expect, it, vi } from "vitest"
 
 import { SITE_TYPES } from "~/constants/siteType"
-import { API_ERROR_CODES, ApiError } from "~/services/apiService/common/errors"
+import { API_ERROR_CODES, ApiError } from "~/services/apiTransport/errors"
 import {
   ACCOUNT_TOKEN_FALLBACK_LOAD_FAILED,
   loadAccountTokenFallbackPricingResponse,
