+++ tests/services/newApiService/newApiService.test.ts
@@ -43,7 +43,7 @@ class MockApiError extends Error {
     this.name = "ApiError"
   }
 }
-vi.mock("~/services/apiService/common/errors", () => ({
+vi.mock("~/services/apiTransport/errors", () => ({
   ApiError: MockApiError,
 }))
